{{cookiecutter.project_slug}}/api/requirements.txt
boto3==1.18.58
botocore==1.21.58
click==8.0.3
Flask==2.0.2
flask-cors==3.0.10
flask-ipban==1.1.5
flask-httpauth==4.4.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Minify==0.32
Flask-SQLAlchemy==2.4.4
flask-talisman==0.8.1
google-cloud-storage==1.42.3
google-cloud-secret-manager==2.7.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
pyjwt==2.1.0
python-dotenv==0.19.0
python-json-logger==2.0.2
requests==2.24.0
retry==0.9.2
selenium==3.141.0
slack-sdk==3.11.2
sphinx==4.1.2
